% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinates_meters.R
\name{add_coordinates_meters}
\alias{add_coordinates_meters}
\title{Add spatial coordinates in the crs in meters}
\usage{
add_coordinates_meters(
  data,
  crs_meters = 3035,
  coordinates_to_write = c("X", "Y")
)
}
\arguments{
\item{data}{data where coordinates are added}

\item{crs_meters}{target crs (must have unit in meters)}

\item{coordinates_to_write}{coordinates to have in the final dataset}
}
\value{
same dataset with added coordinates
}
\description{
Add spatial coordinates in the crs in meters
}
\examples{
\dontrun{
library(AISanalyze)
data("ais")

out <- add_coordinates_meters(ais, crs_meters = 3035, coordinates_to_write = c("X", "Y"))}

}
