% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISextract_all.R
\name{AISextract_all}
\alias{AISextract_all}
\title{Extract all AIS positions in the desired interval of time}
\usage{
AISextract_all(
  data,
  crs_meters = 3035,
  ais_data,
  search_into_radius_m = 50000,
  interval_time_before = 5 * 60,
  interval_time_after = 5 * 60
)
}
\arguments{
\item{data}{Data of interest for the extraction of AIS. Must contain a column: timestamp (number of seconds since January 1, 1970 (the Unix epoch): see https://r-lang.com/how-to-convert-date-to-numeric-format-in-r/ for transformation), and the columns lon (longitude) & lat (latitude). timestamp, lon and lat must be numeric.}

\item{crs_meters}{projection (crs) in 'meters' to use to calculate distance over the study area. Default to 3035 (ETRS89).}

\item{ais_data}{AIS data. Must contain a column: timestamp (number of seconds since January 1, 1970 (the Unix epoch): see https://r-lang.com/how-to-convert-date-to-numeric-format-in-r/ for transformation), and the columns lon (longitude), lat (latitude) and mmsi (Maritime mobile service identity). timestamp, lon and lat must be numeric. The mmsi column is the identifier for the vessels, the values can be replaced by the IMO or another identifier, but the name of the column must be mmsi.}

\item{search_into_radius_m}{radius (kilometer) where the MMSIs are extracted and returned.}

\item{interval_time_before}{Number of seconds before each time from "data" where AIS data are returned. Example: 60 -> all AIS data from 60 seconds before data "time" to the data "time" are returned (calculated for each line of data independantly).}

\item{interval_time_after}{Number of seconds after each time from "data" where AIS data are returned. Example: 60 -> all AIS data from data "time" until 60 seconds after are returned (calculated for each line of data independantly).}
}
\value{
return the input data with the AIS extracted merged in the dataframe: each line of input data is duplicated by timestamp to extract (every "t_gap" number of seconds up to "max_time_diff" number of seconds). All these lines are duplicated for each MMSI present in the area at the moment of the extraction. If no AIS are present in the radius at this moment, the columns dedicated to AIS data are filled with NA, so that no input data and no timestamp to extract is lost.
The output dataframe contains the columns of the input data, the columns of the AIS data (with "ais_" as prefix if the same column is already present in the input data), and the following columns:
\itemize{
\item distance_effort_ais_m: distance (meters) between the data location and the MMSI at this time (filled with NA if no MMSI).
\item timestamp_AIS_to_extract: timestamp for the extraction of the AIS (approximated with "average_at" number of seconds if accelerate = TRUE).
\item diffTime_AIS_extraction_effort: difference (in seconds) between the timestamp to extract (timestamp_AIS_to_extract) and the real data timestamp.
\item datetime_AIS_to_extract: datetime of timestamp_AIS_to_extract.
\item diffTime_AIS_effort: difference, in seconds, between the AIS data and the data timestamp: can be different from diffTime_AIS_extraction_effort due to the parameter "average_at" (that average the timestamp to extract).}
}
\description{
Extract all AIS positions in the desired interval of time
}
\examples{
\dontrun{
data("ais")
data("point_to_extract")

library(dplyr)
library(lubridate)
point_to_extract <- point_to_extract \%>\%
  mutate(timestamp = as.numeric(ymd_hm(datetime)))
ais <- ais \%>\%
  mutate(timestamp = as.numeric(ymd_hms(datetime))) \%>\%
  AIStravel(ais_data = .,
            time_stop = 5*3600,
            mmsi_time_to_order = T,
            return_sf = F,
            return_3035_coords = F) \%>\%
  AISinterpolate_at(data = point_to_extract,
                  ais_data = .,
                  mmsi_time_to_order = T,
                  load_existing_files = T,
                  save_AISinterlate_at = T,
                  overwrite = F,
                  file_AISinterlate_at = "AISinterpolate_at",
                  radius = 200000,
                  time_stop = 5*3600,
                  correct_speed = T,
                  threshold_speed_to_correct = 100,
                  filter_station = T,
                  filter_high_speed = T,
                  parallelize = F)

AISextract_all(data = point_to_extract,
           ais_data = ais,
           search_into_radius_m = 50000,
           interval_time_before = 5 * 60,
           interval_time_after = 5 * 60)}
}
