% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISinterpolate_all.R
\name{AISinterpolate_all}
\alias{AISinterpolate_all}
\title{AISinterpolate_all}
\usage{
AISinterpolate_all(
  ais_data,
  mmsi_time_to_order = T,
  correct_speed = T,
  quantile_station = 0.95,
  threshold_distance_station = 1,
  quantile_high_speed = 0.9,
  threshold_speed_to_correct = 90,
  threshold_high_speed = 90,
  filter_station = T,
  filter_high_speed = T,
  interpolate_station = F,
  interpolate_high_speed = F,
  time_stop = 5 * 60 * 60,
  t_gap = 4,
  spatial_limit = NA,
  on_Land_analysis = T,
  land_sf_polygon = NA
)
}
\arguments{
\item{ais_data}{AIS data. Must contain a column timestamp, lon, lat and mmsi (numeric value). the mmsi column is the identifier for vessel, and values can be replaced by the IMO for example, but the name of the column must be mmsi.}

\item{mmsi_time_to_order}{if MMSI and time are not yet arranged as dplyr::arrange(AIS data, mmsi, timestamp), must be TRUE. We recommand to put it as TRUE by precaution.}

\item{correct_speed}{if speeds of vessel need to be corrected, to remove GPS errors/delay and unrealistic speeds.}

\item{quantile_station}{Quantile of distance by mmsi used to assess if the mmsi is a station or not. We used 0.975 to prevent misinterpretations from GPS errors leading to distance travelled by stations.}

\item{threshold_distance_station}{threshold of distance used to assess if the mmsi is a station.}

\item{quantile_high_speed}{Quantile of speed by mmsi used to assess if the mmsi is a aircraft or not. We used 0.975 to prevent misinterpretations from GPS errors.}

\item{threshold_speed_to_correct}{speeds higher than this threshold are corrected if the mmsi is not an aircraft and if correct_speed = T}

\item{threshold_high_speed}{threshold of speed used to assess if the mmsi is an aircraft.}

\item{filter_station}{if the stations are filtered or not. mmsi_time_to_order}

\item{filter_high_speed}{if the aircraft are filtered or not.}

\item{interpolate_station}{if the stations are interpolated or not.}

\item{interpolate_high_speed}{if the aircraft are interpolated or not.}

\item{time_stop}{number of seconds that looked for interpolation of vessel positions. Interval of time higher than "time_stop" between 2 AIS receptions are considered as a stop of the movement. Filter also AIS data around data timestamp +- time_stop to accelerate the process.}

\item{t_gap}{interval of time (seconds) to which vessels positions are interpolated.}

\item{spatial_limit}{sf polygon object of the area where outside points must be filtered out of the output. Not tested and might lead to few errors.}

\item{on_Land_analysis}{sf polygon object of the countries to study the reliability of GPS positions and interpolations with an analysis of the paths travelled by mmsi on land. Not tested and might lead to few errors.}

\item{land_sf_polygon}{if on_Land_analysis, sf polygon object for countries.}
}
\value{
return AIS data with the columns:
id_ais_data_initial: identifier of the row line in the ais data, ordered, corrected and cleaned. Use for internal computation. For interpolated data, id_ais_data_initial is the same than the next real existing line.
station: if the MMSI is a station or not.
high_speed: if the MMSI is a high speed craft (used for aircraft) or not.
any_NA_speed_kmh: if any of the MMSI point has a value of speed of NA (so distance_travelled or time_travelled has a issue and the MMSI points must be checked). Should not occur.
n_point_mmsi_initial_data: number of point of the MMSI in the initial AIS data, removing firstly the inexisting longitude and latitude points.
id_mmsi_point_initial: identifier for the MMSI point in the ordered, corrected and cleaned AIS data.
speed_kmh_corrected: if the speed of this line has been corrected or not.
interpolated: if this AIS position is an interpolation or not.
}
\description{
AISinterpolate_all
}
\examples{
# to add
}
