% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extend_time.R
\name{data_extend_time}
\alias{data_extend_time}
\title{Title}
\usage{
data_extend_time(
  data,
  max_time_diff = 30 * 60,
  t_gap = 15,
  accelerate = F,
  average_at = 10
)
}
\arguments{
\item{data}{data with a column timestamp (numeric value of time)}

\item{max_time_diff}{number of seconds before the data time, when boat positions are considered/extracted.}

\item{t_gap}{interval of time where vessels positions are extracted, from the data time to "max_time_diff" seconds before.}

\item{accelerate}{TRUE or FALSE: if data times must be averaged within "average_at" seconds, to equlize times where vessels positions are extracted and decreased strongly computation time.}

\item{average_at}{number of seconds where times are averaged if accelerate = TRUE to decrease number of data time to extract. If average_at = 10, data times are averaged in the interval time-5:time+5.}
}
\value{
to add
}
\description{
Title
}
\examples{
to add
}
