% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_ais.R
\name{data_extract_ais}
\alias{data_extract_ais}
\title{Extract vessels data at desired times in a desired radius. Used for AISextract. Use only if sure about this.}
\usage{
data_extract_ais(
  data,
  data_mmsi,
  search_into_radius_m = 20000,
  duplicate_time = T,
  max_time_diff = 0,
  average_at = 0,
  t_gap,
  accelerate = T
)
}
\arguments{
\item{data}{Your data frame, with a column timestamp, lon and lat (numeric value of time, longitude, latitude)}

\item{data_mmsi}{AIS data, with a column timestamp, lon, lat and mmsi (numeric value of time, longitude, latitude, Maritime Mobile Service Identity)}

\item{search_into_radius_m}{radius  in which vessels positions are returned}

\item{duplicate_time}{if vessels positions must be extracted only for data times (TRUE) or every "t_gap" seconds up to max_time_diff before data times.}

\item{max_time_diff}{number of seconds before the data time, when boat positions are considered/extracted.}

\item{average_at}{number of seconds where times are averaged if accelerate = TRUE to decrease number of data time to extract. If average_at = 10, data times are averaged in the interval time-5:time+5.}

\item{t_gap}{interval of time where vessels positions are extracted, from the data time to "max_time_diff" seconds before.}

\item{accelerate}{TRUE or FALSE: if data times must be averaged within "average_at" seconds, to equlize times where vessels positions are extracted and decreased strongly computation time.}
}
\value{
to add
}
\description{
Extract vessels data at desired times in a desired radius. Used for AISextract. Use only if sure about this.
}
\examples{
to add
}
