% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_interpolation_exact_time.R
\name{method_interpolation_exact_time}
\alias{method_interpolation_exact_time}
\title{Interpolate AIS positions}
\usage{
method_interpolation_exact_time(
  ais_data,
  data,
  crs_meters,
  radius,
  timestamp_to_interpolate,
  nb_cores,
  outfile
)
}
\arguments{
\item{ais_data}{AIS data frame containing \code{timestamp}, \code{lon}, \code{lat}, and
\code{mmsi}. \code{timestamp} must be Unix time (seconds since 1970-01-01), while
\code{lon} and \code{lat} must be numeric.}

\item{data}{target timestamps and locations (optional)}

\item{crs_meters}{CRS (in metres) used for distance calculations. Defaults
to EPSG:3035.}

\item{radius}{Optional search radius (m) around each target location.}

\item{timestamp_to_interpolate}{Target timestamps.}

\item{nb_cores}{Number of CPU cores used if \code{type_interpolation = exact_timestamp}.}

\item{outfile}{File used to save logs when \code{type_interpolation = exact_timestamp}.}
}
\value{
The interpolated AIS data with an additional column:
\itemize{
\item \code{interpolated}: Whether the position was interpolated.
}
}
\description{
Interpolates vessel positions at user-defined timestamps.
Interpolation can optionally be restricted to
a given radius within target locations
}
