% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_MMSI_infos.R
\name{MMSI_infos}
\alias{MMSI_infos}
\title{Estimate vessel type and vessel length}
\usage{
MMSI_infos(ais_data, threshold_length = 475, weight_complete_data = 10)
}
\arguments{
\item{ais_data}{AIS data. Must contain a column: timestamp (number of seconds since January 1, 1970 (the Unix epoch): see https://r-lang.com/how-to-convert-date-to-numeric-format-in-r/ for transformation), and the columns lon (longitude), lat (latitude) and mmsi (Maritime mobile service identity). timestamp, lon and lat must be numeric. The mmsi column is the identifier for the vessels, the values can be replaced by the IMO or another identifier, but the name of the column must be mmsi.}

\item{weight_complete_data}{}
}
\description{
Collect information on the type and length of the vessels to identify the (probable) real type and length, among the errors usually present in the AIS data.
}
