% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{point_to_extract}
\alias{point_to_extract}
\title{Example extraction locations}
\format{
A data frame with locations and timestamps of interest
}
\usage{
point_to_extract
}
\description{
Example target locations used in the package examples.
}
\keyword{datasets}
