% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns_data.R
\name{rename_columns_data}
\alias{rename_columns_data}
\title{Check column names of data}
\usage{
rename_columns_data(data)
}
\arguments{
\item{data}{data with target timestamps and locations to extract}
}
\value{
the same dataset, with columns that are renamed if conflict were detected
}
\description{
Check if any column in \code{data} are called \code{mmsi}, and rename it if so to avoid conflict with AIS data.
}
