% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_colums_ais.R
\name{rename_colums_ais}
\alias{rename_colums_ais}
\title{Check column names of AIS data}
\usage{
rename_colums_ais(ais_data, data)
}
\arguments{
\item{ais_data}{AIS data}

\item{data}{data with target timestamps and locations to extract}
}
\value{
the same AIS dataset, with columns that are renamed if conflict were detected
}
\description{
Check if any columns of AIS data are in conflict with those of \code{data}, and rename them if so to avoid conflicts (excepted for \code{mmsi} and \code{timestamp}).
}
