% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIS_all_to_extract.R
\name{AIS_all_to_extract}
\alias{AIS_all_to_extract}
\title{Extract vessels data at desired times in a desired radius}
\usage{
AIS_all_to_extract(
  data,
  ais_data,
  parallelize = F,
  nb_cores = 4,
  outfile = "log.txt",
  run_AIStravel = T,
  save_AIStravel = F,
  file_AIStravel = "AIStravel.rds",
  run_AISinterpolate_at = T,
  save_AISinterlate_at = F,
  file_AISinterlate_at = "AISinterlate_at.rds",
  run_AISextract_perHour = T,
  save_AISextract_perHour = F,
  file_AISextract_perHour = "AISextract.rds",
  return_merged_all_extracted = T,
  time_stop = Inf,
  radius = Inf,
  mmsi_time_to_order = T,
  search_into_radius_m = 20000,
  max_time_diff = 30 * 60,
  duplicate_time = F,
  t_gap = 15,
  average_at = 10,
  average_mmsi_at = 0,
  accelerate = F,
  QUIET = F,
  correct_speed = T,
  quantile_station = 0.95,
  threshold_distance_station = 0,
  quantile_high_speed = 0.9,
  threshold_speed_to_correct = 100,
  threshold_high_speed = 110,
  filter_station = T,
  interpolate_station = F,
  filter_high_speed = T,
  interpolate_high_speed = F,
  spatial_limit = NA,
  on_Land_analysis = F,
  land_sf_polygon = NA,
  overwrite = F
)
}
\arguments{
\item{data}{Data of interest for AIS extraction. Must contain a column "timestamp", "lon" and "lat" (numeric values).}

\item{ais_data}{AIS data. Must contain a column timestamp, lon, lat and mmsi (numeric value). the mmsi column is the identifier for vessel, and values can be replaced by the IMO for example, but the name of the column must be mmsi.}

\item{parallelize}{if the interpolation and extract must be parallelized (required performant computer with increasing AIS dataframe and data timestamp to process.)}

\item{nb_cores}{number of cores to used for the parallelize}

\item{outfile}{file for output if parallelize = T}

\item{run_AIStravel}{if the AIS data must be ran with AIStravel function firstly or not. If already processed with AIStravel, probably not.}

\item{save_AIStravel}{if results from AIStravel must be saved, if run_AIStravel = T.}

\item{file_AIStravel}{if save_AIStravel = T, file where AIS data are saved after AIStravel.}

\item{run_AISinterpolate_at}{if the AIS data must be ran with AISinterpolate_at function firstly or not. If already processed with AISinterpolate_at, probably not.}

\item{save_AISinterlate_at}{if results from AISinterpolate_at must be saved, if run_AISinterpolate_at = T.}

\item{file_AISinterlate_at}{if save_AISinterlate_at = T, file where AIS data are saved after AISinterpolate_at}

\item{run_AISextract_perHour}{if the AIS data must be ran with AISextract, extracting AIS data for data timestamps, up to "max_time_diff".}

\item{save_AISextract_perHour}{if results from AISextract must be saved, if run_AISextract_perHour = T.}

\item{file_AISextract_perHour}{if save_AISextract_perHour = T, file where AIS data are saved after AISextract}

\item{return_merged_all_extracted}{if all extracted AIS data, merged with the data frame of input, must be returned. Otherwise, return NULL. Usefull if results have been saved during the process and the output is not necessary (decrease memory used).}

\item{time_stop}{number of seconds that looked for interpolation of vessel positions. Interval of time higher than "time_stop" between 2 AIS receptions are considered as a stop of the movement. Filter also AIS data around data timestamp +- time_stop to accelerate the process.}

\item{radius}{radius within which AIS data are extracted around the data of input.}

\item{mmsi_time_to_order}{if MMSI and time are not yet arranged as dplyr::arrange(AIS data, mmsi, timestamp), must be TRUE. We recommand to put it as TRUE by precaution.}

\item{search_into_radius_m}{radius in which vessels positions are extracted.}

\item{max_time_diff}{number of seconds before the timestamp of every data timestamp, where boat positions are considered/extracted.}

\item{duplicate_time}{if vessels positions must be extracted only for data timestamp (TRUE), or every "t_gap" seconds, up to "max_time_diff" before the data timestamps.}

\item{t_gap}{interval of time into which vessels positions are extracted, from the data timestamp up to "max_time_diff" seconds before. This defines also the time interval where boat are considered for the extraction.}

\item{average_at}{if accelerate = TRUE, average the data timestamps to +- average_at to decrease the number of data timestamp to process. This defines also the time interval where boat are considered for the extraction.}

\item{average_mmsi_at}{number of seconds where positions of mmsi are averaged for extraction. Less useful than average_at which average the data timestamps to process.}

\item{accelerate}{TRUE or FALSE: if data timestamps must be averaged at "average_at" seconds to decrease the number of data timestamp to process and  strongly decrease the computation time.}

\item{QUIET}{if iterations are printed, either in the console if parallelize = F, or in the file "outfile" if parallelize = T.}

\item{correct_speed}{if speeds of vessel need to be corrected, to remove GPS errors/delay and unrealistic speeds.}

\item{quantile_station}{Quantile of distance by mmsi used to assess if the mmsi is a station or not. We used 0.975 to prevent misinterpretations from GPS errors leading to distance travelled by stations.}

\item{threshold_distance_station}{threshold of distance used to assess if the mmsi is a station.}

\item{quantile_high_speed}{Quantile of speed by mmsi used to assess if the mmsi is a aircraft or not. We used 0.975 to prevent misinterpretations from GPS errors.}

\item{threshold_speed_to_correct}{speeds higher than this threshold are corrected if the mmsi is not an aircraft and if correct_speed = T}

\item{threshold_high_speed}{threshold of speed used to assess if the mmsi is an aircraft.}

\item{filter_station}{if the stations are filtered or not.}

\item{interpolate_station}{if the stations are interpolated or not.}

\item{filter_high_speed}{if the aircraft are filtered or not.}

\item{interpolate_high_speed}{if the aircraft are interpolated or not.}

\item{spatial_limit}{sf polygon object of the area where outside points must be filtered out of the output. Not tested and might lead to few errors.}

\item{on_Land_analysis}{sf polygon object of the countries to study the reliability of GPS positions and interpolations with an analysis of the paths travelled by mmsi on land. Not tested and might lead to few errors.}

\item{land_sf_polygon}{if on_Land_analysis, sf polygon object for countries.}

\item{overwrite}{if any file must be written (save_AIStravel, save_AISinterlate_at, save_AISextract_perHour), if file are overwritten. Otherwise load the files if files are already existing and if the file is needed in the function.}
}
\value{
to add
}
\description{
Extract vessels data at desired times in a desired radius
}
\examples{
# to add
}
