% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISanalyze-package.R
\docType{package}
\name{AISanalyze}
\alias{AISanalyze-package}
\alias{AISanalyze}
\title{AISanalyze}
\description{
Tools for analysing, cleaning, interpolating and extracting Automatic
Identification System (AIS) vessel data.
}
\details{
\subsection{Main functions:}{

\strong{Travel metrics}
\itemize{
\item \code{\link[=AIStravel]{AIStravel()}}
\item \code{\link[=AIScorrect_speed]{AIScorrect_speed()}}
}

\strong{Vessel characteristics}
\itemize{
\item \code{\link[=AISinfos]{AISinfos()}}
\item \code{\link[=AISidentify_stations_aircraft]{AISidentify_stations_aircraft()}}
}

\strong{Interpolation}
\itemize{
\item \code{\link[=AISinterpolate]{AISinterpolate()}}
}

\strong{Extraction}
\itemize{
\item \code{\link[=AISextract]{AISextract()}}
}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/remip48/AISanalyze}
  \item \url{https://remip48.github.io/AISanalyze/}
  \item Report bugs at \url{https://github.com/remip48/AISanalyze/issues}
}

}
\author{
\strong{Maintainer}: R<U+00E9>mi Pigeault \email{remi.pigeault@tiho-hannover.de} (\href{https://orcid.org/0000-0003-2714-5122}{ORCID})

}
\keyword{internal}
