% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISanalyze-package.R
\docType{package}
\name{AISanalyze}
\alias{AISanalyze-package}
\alias{AISanalyze}
\title{AISanalyze}
\description{
Tools for analysing AIS data.
}
\details{
Main functions:
\itemize{
\item \code{\link[=AIStravel]{AIStravel()}}
\item \code{\link[=AIScorrect_speed]{AIScorrect_speed()}}
\item \code{\link[=AISidentify_stations_aircraft]{AISidentify_stations_aircraft()}}
\item \code{\link[=AISinterpolate]{AISinterpolate()}}
\item \code{\link[=AISextract]{AISextract()}}
\item \code{\link[=AISinfos]{AISinfos()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/remip48/AISanalyze}
  \item Report bugs at \url{https://github.com/remip48/AISanalyze/issues}
}

}
\author{
\strong{Maintainer}: Rémi Pigeault \email{remi.pigeault@tiho-hannover.de}

Authors:
\itemize{
  \item Rémi Pigeault \email{remi.pigeault@tiho-hannover.de}
}

}
