% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIScorrect_speed.R
\name{AIScorrect_speed}
\alias{AIScorrect_speed}
\title{Correct GPS errors in AIS tracks}
\usage{
AIScorrect_speed(
  ais_data,
  crs_meters = 3035,
  threshold_speed_to_correct = 100,
  threshold_speed_to_correct_function = NULL,
  nb_cores = 1,
  outfile = tempfile()
)
}
\arguments{
\item{ais_data}{AIS data frame containing \code{timestamp}, \code{lon}, \code{lat}, and
\code{mmsi}. \code{timestamp} must be Unix time (seconds since 1970-01-01), while
\code{lon} and \code{lat} must be numeric. Another vessel identifier may be used if
the column is named \code{mmsi}.}

\item{crs_meters}{CRS (metres) used to calculate travelled distances.
Defaults to EPSG:3035.}

\item{threshold_speed_to_correct}{Speed threshold (km/h) above which
observations are corrected.}

\item{threshold_speed_to_correct_function}{a function to estimate vessel-specific
speed thresholds. The function can use all columns of ais_data as argument
(such as \code{speed_kmh}, \code{distance_travelled} or \code{time_travelled}). If set to
\code{NULL}, an internal function is used (see \code{details}).}

\item{nb_cores}{Number of CPU cores used.}

\item{outfile}{File used to save logs.}
}
\value{
The input data with corrected travel metrics and the following
columns:
\itemize{
\item \code{speed_kmh_corrected}: Whether the speed was corrected.
\item \code{time_travelled}: Travel time (s).
\item \code{distance_travelled}: Travelled distance (m).
\item \code{speed_kmh}: Vessel speed (km/h).
}
}
\description{
Detects and corrects GPS errors and delayed AIS messages that generate
unrealistic vessel speeds, travelled distances, and travel times.
Note: For consecutive GPS errors, only the first point is removed to
avoid overcorrection.
}
\details{
When \code{threshold_speed_to_correct_function} is set to \code{NULL}, a vessel-specific
speed threshold is estimated from the observed vessel \code{speed_kmh}. The
threshold is calculated as:

\deqn{
T = 15 + \operatorname{median}(v_{>1}) +
5 \times \operatorname{SD}(v_{>1,\;v<Q_{0.75}})
}{
T = 15 + median(v > 1) +
5 * SD(v > 1 and v < Q_{0.75})
}

where \eqn{v_{>1}} represents vessel speeds greater than 1 km/h, and
\eqn{v_{>1,\;v<Q_{0.75}}} represents vessel speeds between 1 km/h and the
75th percentile. The equation is designed to capture the typical
travelling speed of the vessel while accounting for variation in its
observed travelling speeds.
}
\examples{
library(AISanalyze)
data("ais")

# Define the Unix time (seconds since 1970-01-01)
ais$timestamp <- as.numeric(lubridate::ymd_hms(ais$datetime))

# calculate the travelled distance, time, and speed:
ais <- AIStravel(ais_data = ais)

# Correct speed:
out <- AIScorrect_speed(ais_data = ais,
                 crs_meters = 3035,
                 threshold_speed_to_correct = 100)
}
