% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIScorrect_speed.R
\name{AIScorrect_speed}
\alias{AIScorrect_speed}
\title{Correct GPS errors in AIS tracks}
\usage{
AIScorrect_speed(
  ais_data,
  crs_meters = 3035,
  threshold_speed_to_correct = 100,
  threshold_speed_to_correct_expr = function(speed_kmh) {
     return(15 +
    stats::median(speed_kmh[speed_kmh > 1], na.rm = T) + 5 *
    stats::sd(speed_kmh[speed_kmh > 1 & speed_kmh < stats::quantile(speed_kmh[speed_kmh >
    1], 0.75)]))
 }
)
}
\arguments{
\item{ais_data}{AIS data frame containing \code{timestamp}, \code{lon}, \code{lat}, and
\code{mmsi}. \code{timestamp} must be Unix time (seconds since 1970-01-01), while
\code{lon} and \code{lat} must be numeric. Another vessel identifier may be used if
the column is named \code{mmsi}.}

\item{crs_meters}{CRS (metres) used to calculate travelled distances.
Defaults to EPSG:3035.}

\item{threshold_speed_to_correct}{Speed threshold (km/h) above which
observations are corrected.}

\item{threshold_speed_to_correct_expr}{Function returning a vessel-specific
speed threshold from \code{speed_kmh}.}
}
\value{
The input data with corrected travel metrics and the following
columns:
\itemize{
\item \code{speed_kmh_corrected}: Whether the speed was corrected.
\item \code{time_travelled}: Travel time (s).
\item \code{distance_travelled}: Travelled distance (m).
\item \code{speed_kmh}: Vessel speed (km/h).
}
}
\description{
Detects and corrects GPS errors and delayed AIS messages that generate
unrealistic vessel speeds, travelled distances, and travel times.
}
\examples{
\dontrun{
library(AISanalyze)
data("ais")

ais <- ais \%>\%
  mutate(timestamp = as.numeric(ymd_hms(datetime))) \%>\%
  AIStravel(ais_data = .)

out <- AIScorrect_speed(ais_data = ais,
                 crs_meters = 3035,
                 threshold_speed_to_correct = 100)}
}
