% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02b_AIScorrect_speed.R
\name{AIScorrect_speed}
\alias{AIScorrect_speed}
\title{Correct GPS errors and delays from AIS data}
\usage{
AIScorrect_speed(
  ais_data,
  mmsi_time_to_order = T,
  correct_high_speed_craft = F,
  threshold_speed_to_correct = 100,
  threshold_speed_to_correct_expr = function(speed_kmh) {
    
    return((median(speed_kmh[speed_kmh > 0], na.rm = T) + sd(speed_kmh[speed_kmh > 0 &
    speed_kmh < quantile(speed_kmh, 0.75, na.rm = T)]) * 2.5 + 15))
 },
  time_stop = 5 * 60 * 60,
  nb_iteration = 1,
  nb_iteration_successive_strange = 10
)
}
\arguments{
\item{ais_data}{AIS data. Must contain a column: timestamp (number of seconds since January 1, 1970 (the Unix epoch): see https://r-lang.com/how-to-convert-date-to-numeric-format-in-r/ for transformation), and the columns lon (longitude), lat (latitude) and mmsi (Maritime mobile service identity). timestamp, lon and lat must be numeric. The mmsi column is the identifier for the vessels, the values can be replaced by the IMO or another identifier, but the name of the column must be mmsi.}

\item{mmsi_time_to_order}{if MMSI and timestamps are not yet arranged as dplyr::arrange(AIS data, mmsi, timestamp), must be TRUE. We recommand to put it as TRUE by precaution. Important to prevent large errors.}

\item{correct_high_speed_craft}{if speeds from high speed craft (specially used for aircraft) must be corrected as well.}

\item{threshold_speed_to_correct}{speeds higher than this threshold are corrected if the mmsi is not an aircraft and if correct_speed = T}

\item{threshold_speed_to_correct_expr}{expression (function having "speed_kmh" as unique parameter) to determine another threshold correcting GPS errors and delays. This expression is ran for each MMSI individually, allowing to identify unrealistic speeds based on the mean of the vessel speed, median, standard deviation or other functions.}

\item{time_stop}{number of seconds before and after the AIS signal were the vessel track is not calculated/interpolated anymore if there is not another AIS signal meanwhile. Filter also AIS data too long before and after that are not of interest, to accelerate a lot the process.}

\item{nb_iteration}{number of iteration re-estimating the GPS errors and delays. Low number (1 as default) is prefered.}

\item{nb_iteration_successive_strange}{number of iterations where GPS delays are re-estimated. 10 as default. Tests showed that this number of iteration can go up to very high value for different reasonable (real delays) and unreasonable reasons (GPS error or delay impossible to correct with this code only. Exemple: positions of stationary ships are several shifted of several hundreds of kilometers for unknown reason and for several hours. In this code, we did not find a solution to identify automatically wether the initial position in this case is the real position, or the shifted and erroneous position. Such that this error can not be solved at the moment).}
}
\value{
Return the AIS data with identified GPS error or GPS delay removed. The vessel speeds are corrected, as well as the distance and time travelled. Contains columns:
speed_kmh_corrected: if TRUE, the speed of the line has been corrected.
time_travelled: number of seconds since the last reception or interpolation of an AIS signal (0 if first reception).
distance_travelled: distance travelled (meters) since the last reception or interpolation of an AIS signal (0 if first reception).
speed_kmh: speed (km/h) of the vessels since the last reception or interpolation of an AIS signal.
station: if the MMSI is a station or not.
high_speed: if the MMSI is a high speed craft (used for aircraft) or not.
}
\description{
Correct the errors due to GPS delays or errors and leading to erroneous speeds (km/h), distance or time travelled by MMSI (vessel). The corrections are made according to different input parameters.
}
