% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISidentify_stations_aircraft.R
\name{AISidentify_stations_aircraft}
\alias{AISidentify_stations_aircraft}
\title{Identify AIS base stations and high-speed craft}
\usage{
AISidentify_stations_aircraft(ais_data, crs_meters = 3035)
}
\arguments{
\item{ais_data}{AIS data frame containing \code{timestamp}, \code{lon}, \code{lat}, and
\code{mmsi}. \code{timestamp} must be Unix time (seconds since 1970-01-01), while
\code{lon} and \code{lat} must be numeric. Another vessel identifier may be used if
the column is named \code{mmsi}.}

\item{crs_meters}{CRS (metres) used to calculate distances. Defaults to
EPSG:3035.}
}
\value{
The input AIS data with the following additional columns:
\itemize{
\item \code{station}: Whether the MMSI is classified as a base station.
\item \code{high_speed}: Whether the MMSI is classified as a high-speed craft.
\item \code{n_point_mmsi_initial_data}: Number of valid AIS positions for the MMSI.
\item \code{id_mmsi_point_initial}: Sequential identifier of each AIS position.
}
}
\description{
Identify AIS base stations and high-speed craft
}
\examples{
\dontrun{
library(AISanalyze)
data("ais")

ais <- ais |>
  mutate(timestamp = as.numeric(ymd_hms(datetime))) |>
  AIStravel(.)

out <- AISidentify_stations_aircraft(ais_data = ais)}
}
