% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISinterpolate.R
\name{AISinterpolate}
\alias{AISinterpolate}
\title{Interpolate AIS positions}
\usage{
AISinterpolate(
  ais_data,
  type_interpolation,
  maximum_gap_seconds,
  exact_timestamp = list(timestamp_to_interpolate, locations_of_interest, radius),
  crs_meters = 3035,
  nb_cores = 1,
  outfile = tempfile()
)
}
\arguments{
\item{ais_data}{AIS data frame containing \code{timestamp}, \code{lon}, \code{lat}, and
\code{mmsi}. \code{timestamp} must be Unix time (seconds since 1970-01-01), while
\code{lon} and \code{lat} must be numeric.}

\item{type_interpolation}{Interpolation mode: \code{"maximum_gap_seconds"} or
\code{"exact_timestamp"}.}

\item{maximum_gap_seconds}{used when
\code{type_interpolation = "maximum_gap_seconds"}: threshold above which
AIS signals are interpolated.}

\item{exact_timestamp}{List used when
\code{type_interpolation = "exact_timestamp"}, containing:
\itemize{
\item \code{timestamp_to_interpolate}
\item \code{locations_of_interest}: (optional) data frame with \code{lon} and \code{lat}
columns corresponding to each \code{timestamp_to_interpolate}
\item \code{radius}: (optional) a search radius (m) around target locations
}}

\item{crs_meters}{CRS (metres) used to calculate distances
in the study area (defaults to EPSG:3035, Europe).
Tip: use \code{suggest_crs} function (\code{crsuggest} package) to find a suitable
CRS for your study area.}

\item{nb_cores}{Number of CPU cores used.}

\item{outfile}{File used to save logs.}
}
\value{
The interpolated AIS data with an additional column:
\itemize{
\item \code{interpolated}: Whether the position was interpolated.
}
}
\description{
Interpolates vessel positions either: (depending on \code{type_interpolation})
\itemize{
\item to ensure time intervals do not exceed a specified maximum
(\verb{= maximum_gap_seconds}).
\item at user-defined timestamps (\verb{= exact_timestamp}). Interpolation
can optionally be restricted to specific regions to reduce computation
time with \code{locations_of_interest} and \code{radius} arguments.
}
}
\examples{
data("ais")
data("point_to_extract")

# use only a sample for the example:
ais <- ais[20000:30000, ]

# Define the Unix time (seconds since 1970-01-01)
point_to_extract$timestamp <- as.numeric(lubridate::ymd_hm(point_to_extract$datetime))
ais$timestamp <- as.numeric(lubridate::ymd_hms(ais$datetime))

# calculate the travelled distance, time, and speed:
ais <- AIStravel(ais_data = ais, crs_meters = 3035)

# Interpolate all AIS signals further than > 120 seconds:
out <- AISinterpolate(ais_data = ais,
                      type_interpolation = "maximum_gap_seconds",
                      maximum_gap_seconds = 120, ## Alternatively, you can
                      ## interpolate at target timestamps with:
                      # exact_timestamp = list(
                      #      timestamp_to_interpolate = point_to_extract$timestamp,
                      #      locations_of_interest = data.frame(lon = point_to_extract$lon,
                      #                                         lat = point_to_extract$lat),
                      #      radius = 200000),
                      crs_meters = 3035)
}
