% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISinterpolate.R
\name{AISinterpolate}
\alias{AISinterpolate}
\title{Interpolate AIS positions}
\usage{
AISinterpolate(
  ais_data,
  type_interpolation,
  maximum_gap_seconds,
  exact_timestamp = list(timestamp_to_interpolate, locations_of_interest, radius),
  crs_meters = 3035,
  nb_cores = 1,
  outfile = "log.txt"
)
}
\arguments{
\item{ais_data}{AIS data frame containing \code{timestamp}, \code{lon}, \code{lat}, and
\code{mmsi}. \code{timestamp} must be Unix time (seconds since 1970-01-01), while
\code{lon} and \code{lat} must be numeric.}

\item{type_interpolation}{Interpolation mode: \code{"maximum_gap_seconds"} or
\code{"exact_timestamp"}.}

\item{maximum_gap_seconds}{used when
\code{type_interpolation = "maximum_gap_seconds"}: threshold above which
AIS signals are interpolated.}

\item{exact_timestamp}{List used when
\code{type_interpolation = "exact_timestamp"}, containing:
\itemize{
\item \code{timestamp_to_interpolate}: Target timestamps.
\item \code{locations_of_interest}: Optional data frame with \code{lon} and \code{lat}
columns corresponding to each \code{timestamp_to_interpolate}.
\item \code{radius}: Optional search radius (m) around each target location.
}}

\item{crs_meters}{CRS (in metres) used for distance calculations. Defaults
to EPSG:3035.}

\item{nb_cores}{Number of CPU cores used if \code{type_interpolation = exact_timestamp}.}

\item{outfile}{File used to save logs when \code{type_interpolation = exact_timestamp}.}
}
\value{
The interpolated AIS data with an additional column:
\itemize{
\item \code{interpolated}: Whether the position was interpolated.
}
}
\description{
Interpolates vessel positions either: (depending on \code{type_interpolation})
\itemize{
\item to ensure time intervals do not exceed a specified maximum
(\code{maximum_gap_seconds}).
\item at user-defined timestamps (\code{exact_timestamp}). Interpolation
can optionally be restricted to a given radius within
target locations to reduce computation time.
}
}
\examples{
\dontrun{
library(AISanalyze)
data("ais")
data("point_to_extract")

point_to_extract <- point_to_extract \%>\%
  mutate(timestamp = as.numeric(ymd_hm(datetime)))

ais <- ais \%>\%
  mutate(timestamp = as.numeric(ymd_hms(datetime))) \%>\%
  AIStravel(ais_data = .)

# to interpolate all vessel locations separated by > 60 seconds
out <- AISinterpolate(ais_data = ais,
               type_interpolation = "maximum_gap_seconds",
               maximum_gap_seconds = 60,
               crs_meters = 3035)

# to interpolate all vessel locations at exact timestamps,
# within a radius of 200 000 meters around
# target locations
out <- AISinterpolate(ais_data = ais,
           type_interpolation = "exact_timestamp",
           exact_timestamp = list(
               timestamp_to_interpolate = point_to_extract$timestamp,
               locations_of_interest = data.frame(lon = point_to_extract$lon,
                                                 lat = point_to_extract$lat),
               radius = 200000),
           crs_meters = 3035)
           }
}
