% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_AISinterpolate_all.R
\name{AISinterpolate_all}
\alias{AISinterpolate_all}
\title{to interpolate AIS data for each mmsi}
\usage{
AISinterpolate_all(
  ais_data,
  mmsi_time_to_order = T,
  t_gap = 30,
  time_stop = 5 * 60 * 60,
  correct_speed = T,
  threshold_speed_to_correct = 100,
  filter_station = T,
  filter_high_speed = T,
  quantile_station = 0.975,
  threshold_distance_station = 10,
  quantile_high_speed = 0.97,
  threshold_high_speed = 110,
  interpolate_station = F,
  interpolate_high_speed = F
)
}
\arguments{
\item{ais_data}{AIS data. Must contain a column timestamp, lon, lat and mmsi (numeric value). the mmsi column is the identifier for vessel, and values can be replaced by the IMO for example, but the name of the column must be mmsi.}

\item{mmsi_time_to_order}{if MMSI and time are not yet arranged as dplyr::arrange(AIS data, mmsi, timestamp), must be TRUE. We recommand to put it as TRUE by precaution.}

\item{t_gap}{interval of time (seconds) to which vessels positions are interpolated.}

\item{time_stop}{number of seconds around the AIS reception considered for interpolation of vessel positions. Interval of time higher than "time_stop" between 2 AIS receptions are considered as a stop of the movement. Filter also AIS data around data timestamp +- time_stop to accelerate the process.}

\item{correct_speed}{if speeds of vessel need to be corrected, to remove GPS errors/delay and unrealistic speeds.}

\item{threshold_speed_to_correct}{speeds higher than this threshold are corrected if the mmsi is not an aircraft and if correct_speed = T}

\item{filter_station}{if the stations are filtered or not. mmsi_time_to_order}

\item{filter_high_speed}{if the aircraft are filtered or not.}

\item{quantile_station}{Quantile of distance by mmsi used to assess if the mmsi is a station or not. We used 0.975 to prevent misinterpretations from GPS errors leading to distance travelled by stations.}

\item{threshold_distance_station}{threshold of distance used to assess if the mmsi is a station.}

\item{quantile_high_speed}{Quantile of speed by mmsi used to assess if the mmsi is a aircraft or not. We used 0.975 to prevent misinterpretations from GPS errors.}

\item{threshold_high_speed}{threshold of speed used to assess if the mmsi is an aircraft.}

\item{interpolate_station}{if the stations are interpolated or not.}

\item{interpolate_high_speed}{if the aircraft are interpolated or not.}
}
\value{
return AIS data with the columns:
id_ais_data_initial: identifier of the row line in the ais data, ordered, corrected and cleaned. Use for internal computation. For interpolated data, id_ais_data_initial is the same than the next real existing line.
station: if the MMSI is a station or not.
high_speed: if the MMSI is a high speed craft (used for aircraft) or not.
any_NA_speed_kmh: if any of the MMSI point has a value of speed of NA (so distance_travelled or time_travelled has a issue and the MMSI points must be checked). Should not occur.
n_point_mmsi_initial_data: number of point of the MMSI in the initial AIS data, removing firstly the inexisting longitude and latitude points.
id_mmsi_point_initial: identifier for the MMSI point in the ordered, corrected and cleaned AIS data.
speed_kmh_corrected: if the speed of this line has been corrected or not.
interpolated: if this AIS position is an interpolation or not.
}
\description{
to interpolate AIS data for each mmsi
}
\examples{
# to add
}
