% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AISinterpolate_at.R
\name{AISinterpolate_at}
\alias{AISinterpolate_at}
\title{Interpolate vessel positions at the time desired}
\usage{
AISinterpolate_at(
  ais_data,
  data_to_interpolate,
  overwrite,
  file_AISinterlate_at,
  mmsi_time_to_order = T,
  average_mmsi_at = 0,
  parallelize = F,
  nb_cores = 4,
  outfile = "log.txt",
  QUIET = F,
  radius = Inf,
  correct_speed = T,
  quantile_station = 0.95,
  threshold_distance_station = 1,
  quantile_high_speed = 0.9,
  threshold_speed_to_correct = 90,
  threshold_high_speed = 90,
  filter_station = T,
  filter_high_speed = T,
  interpolate_station = F,
  interpolate_high_speed = F,
  time_stop = Inf,
  spatial_limit = NA,
  on_Land_analysis = T,
  land_sf_polygon = NA,
  return_all = F
)
}
\arguments{
\item{ais_data}{AIS data, with a column timestamp, lon, lat and mmsi (numeric value of time, longitude, latitude, Maritime Mobile Service Identity). Must be an output of AIStravel and contain distance_travelled, time_travelled, speed_kmh columns.}

\item{data_to_interpolate}{data times where vessel positions must be interpolate. Contain timestamp, lon and latitude columns}

\item{mmsi_time_to_order}{if mmsi & time column must be ordered, if FALSE, is already ordered by MMSI and then by timestamp (arrange(mmsi, timestamp)).}

\item{QUIET}{to print steps along the process or not.}

\item{radius}{radius around data where interpolations are performed. Decrease the radius to decrease computation time.}

\item{correct_speed}{if points with high and unrealistic speed must be removed and speed at this location corrected from previous and next point.}

\item{quantile_station}{if a mmsi have a percentage of data higher than this quantile, with distance_travelled < threshold_distance_station, this mmsi is designed as station (station = TRUE)}

\item{threshold_distance_station}{define threshold of distance, in meters, identifying stations among mmsi.}

\item{quantile_high_speed}{if a mmsi have a percentage of data higher than this quantile, with speed_kmh > threshold_high_speed, this mmsi is designed as high_speed (possible aircraft) (high_speed = TRUE)}

\item{threshold_speed_to_correct}{define unrealistic speed for correct_speed argument.}

\item{threshold_high_speed}{define threshold of speed, in km/h, identifying high_speed among mmsi.}

\item{filter_station}{if stations must be filtered.}

\item{filter_high_speed}{if high_speed mmsi must be filtered.}

\item{interpolate_station}{f stations must be interpolated.}

\item{interpolate_high_speed}{if high_speed mmsi must be interpolated.}

\item{time_stop}{time defining the maximum time interval where a vessel position is interpolated.}

\item{spatial_limit}{sf polygon defining the spatial limit of interpolation, otherwise let NA.}

\item{on_Land_analysis}{if a spatial analysis must be conducted to estimate time and distance of vessels and on interpolation on land (to determine AIS and interpolation error). Make computation longer.}

\item{land_sf_polygon}{if on_Land_analysis = TRUE, sf polygon containing the land.}

\item{return_all}{if all ais data must be returned, with interpolated positions, or only positions of the vessels at the time desired (smaller dataset).}
}
\value{
to add
}
\description{
Interpolate vessel positions at the time desired
}
\examples{
# to add
}
