% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIStravel.R
\name{AIStravel}
\alias{AIStravel}
\title{Title}
\usage{
AIStravel(AIS, time_stop = Inf)
}
\arguments{
\item{AIS}{AIS data, with a column timestamp, lon, lat and mmsi (numeric value of time, longitude, latitude, Maritime Mobile Service Identity).}

\item{time_stop}{time defining the maximum time interval where a vessel position is interpolated.}
}
\value{
to add
}
\description{
Title
}
\examples{
to add
}
