% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIStravel.R
\name{AIStravel}
\alias{AIStravel}
\title{AIStravel}
\usage{
AIStravel(
  AIS,
  time_stop = Inf,
  mmsi_time_to_order = T,
  return_sf = F,
  return_3035_coords = F
)
}
\arguments{
\item{time_stop}{number of seconds that looked for interpolation of vessel positions. Interval of time higher than "time_stop" between 2 AIS receptions are considered as a stop of the movement. Filter also AIS data around data timestamp +- time_stop to accelerate the process.}

\item{mmsi_time_to_order}{if MMSI and time are not yet arranged as dplyr::arrange(AIS data, mmsi, timestamp), must be TRUE. We recommand to put it as TRUE by precaution.}

\item{return_sf}{if the output must be a sf object or a data frame with coordinates as column variable.}

\item{return_3035_coords}{if ETRS3035 coordinates must be returned as X and Y column variables in the output data.}

\item{AOS}{AIS data. Must contain a column timestamp, lon, lat and mmsi (numeric value). the mmsi column is the identifier for vessel, and values can be replaced by the IMO for example, but the name of the column must be mmsi.}
}
\value{
return the AIS data with the speed calculated based on the distance and time travelled since last AIS reception. Contains the columns:
time_travelled: number of seconds since the last reception of an AIS signal (0 if first reception).
distance_travelled:  distance travelled (meters) since the last reception of an AIS signal (0 if first reception).
speed_kmh: speed (km/h) of the vessel since the last reception of an AIS signal.
X and Y columns if return_3035_coords = T (ETRS3035 coordinates)
}
\description{
AIStravel
}
\examples{
# to add
}
