% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AIStravel.R
\name{AIStravel}
\alias{AIStravel}
\title{Estimate distance, time and speed (km/h) travelled by vessels}
\usage{
AIStravel(
  ais_data,
  time_stop = 5 * 60 * 60,
  mmsi_time_to_order = T,
  return_sf = F,
  return_3035_coords = F
)
}
\arguments{
\item{ais_data}{AIS data. Must contain a column: timestamp (number of seconds since January 1, 1970 (the Unix epoch): see https://r-lang.com/how-to-convert-date-to-numeric-format-in-r/ for transformation), and the columns lon (longitude), lat (latitude) and mmsi (Maritime mobile service identity). timestamp, lon and lat must be numeric. The mmsi column is the identifier for the vessels, the values can be replaced by the IMO or another identifier, but the name of the column must be mmsi.}

\item{time_stop}{number of seconds before and after the AIS signal were the vessel track is not calculated/interpolated anymore if there is not another AIS signal meanwhile. Filter also AIS data too long before and after that are not of interest, to accelerate a lot the process.}

\item{mmsi_time_to_order}{if MMSI and timestamps are not yet arranged as dplyr::arrange(AIS data, mmsi, timestamp), must be TRUE. We recommand to put it as TRUE by precaution. Important to prevent large errors.}

\item{return_sf}{if TRUE, return an sf object. If FALSE a data frame with coordinates as column variable.}

\item{return_3035_coords}{if TRUE, return the ETRS3035 coordinates as X and Y columns in the output.}
}
\value{
return the AIS data with the distance, time travelled and the speed of the vessels since the last AIS reception. Contains the columns:
\itemize{
\item time_travelled: number of seconds since the last reception of an AIS signal (0 if first reception).
\item distance_travelled:  distance travelled (meters) since the last reception of an AIS signal (0 if first reception).
\item speed_kmh: speed (km/h) of the vessels since the last reception of an AIS signal.
\item X and Y columns (ETRS3035 coordinates) if return_3035_coords = T.
}
}
\description{
Calculate the distance (meters), time (seconds) and speed (km/h) travelled by each MMSI (vessel) between each AIS reception.
}
\examples{
examples/AIStravel.R
}
