% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02a_DATAextend_time.R
\name{DATAextend_time}
\alias{DATAextend_time}
\title{Extend timestamps of data to past timestamps}
\usage{
DATAextend_time(
  data,
  max_time_diff = 1 * 60 * 60,
  t_gap = 30,
  accelerate = T,
  average_at = 30
)
}
\arguments{
\item{data}{Data of interest for the extraction of AIS. Must contain a column: timestamp (number of seconds since January 1, 1970 (the Unix epoch): see https://r-lang.com/how-to-convert-date-to-numeric-format-in-r/ for transformation), and the columns lon (longitude) & lat (latitude). timestamp, lon and lat must be numeric.}

\item{max_time_diff}{if duplicate_time = TRUE, extend (and duplicate) the data to past timestamps, to investigate the past presence of vessels at the data locations: extend the data timestamps up to "max_time_diff" number of seconds before the timestamps, by steps of "t_gap" number of seconds.}

\item{t_gap}{see "max_time_diff". Is also used as the number of seconds before and after the data timestamps where vessels are considered for extraction, in addition to "average_at" parameter (otherwise other AIS data are filtered out).}

\item{accelerate}{if TRUE, data timestamps are averaged at "average_at" seconds to decrease the number of data timestamp to process and  strongly decrease the computation time.}

\item{average_at}{if accelerate = TRUE, the data timestamps are approximated to within to "average_at" number of seconds. This, to decrease the number of data timestamps to process. Necessary for large data timestamps to extract. Is also used as the number of seconds before and after the data timestamps where vessels are considered for extraction, in addition to "t_gap" parameter (otherwise other AIS data are filtered out).}
}
\value{
The data frame with the lines duplicated for each timestamps, from the data timestamps up to "max_time_diff" by steps of "t_gap". Contains the columns:
\itemize{
\item timestamp_AIS_to_extract: timestamp for the extraction of the AIS (approximated with "average_at" number of seconds if accelerate = TRUE).
\item diffTime_AIS_extraction_effort: difference (in seconds) between the timestamp to extract (timestamp_AIS_to_extract) and the real data timestamp.
\item datetime_AIS_to_extract: datetime of timestamp_AIS_to_extract.
}
}
\description{
Extend timestamps of data to past timestamps, by duplicating the data lines (except timestamp, which goes from the timestamp of the data to "max_time_diff" seconds before, by steps of "t_gap"). Used in internal functions to investigate the AIS presence before the timestamp of the data.
}
\examples{
\dontrun{
DATAextend_time(data,
max_time_diff = 1 * 60 * 60,
t_gap = 30,
accelerate = T,
average_at = 30)}
}
