% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_MMSI_infos.R
\name{MMSI_infos}
\alias{MMSI_infos}
\title{Estimate vessel type and vessel length}
\usage{
MMSI_infos(ais_data, threshold_length = 475, weight_complete_data = 10)
}
\arguments{
\item{ais_data}{AIS data. Must contain the columns shiptype, length and mmsi (Maritime mobile service identity). The mmsi column is the identifier for the vessels, the values can be replaced by the IMO or another identifier, but the name of the column must be mmsi.}

\item{threshold_length}{threshold (meters) above which the length is considered as unrealistic (and set as NA).}

\item{weight_complete_data}{weigth for lines having both the information on length and on type: complete lines are supposed to be more reliable (10 by default for the weight).}
}
\value{
A data frame with one MMSI by row, and containing the columns:
\itemize{
\item npoint: total number of AIS receptions for this MMSI.
\item npoint_all_lengths: number of AIS reception having a value in the column length for this MMSI.
\item npoint_all_lengths_weighted: number of AIS reception (weighted by "weight_complete_data") having a value in the column length for this MMSI.
\item all_lengths: all the values of length found in the AIS data for this MMSI.
\item n_length: number of values of length found in the AIS data for this MMSI.
\item length: estimated length of the MMSI, based on the weighted number of points of each length. The length having the greatest score is considered as the real length.
\item npoint_length: number of AIS reception having the value "length" (estimated real length) for this MMSI.
\item npoint_length_weighted: number of AIS reception (weighted by "weight_complete_data") having the value "length" (estimated real length) for this MMSI.
\item npoint_all_type: number of AIS reception having a value in the column shiptype for this MMSI.
\item npoint_all_type_weighted: number of AIS reception (weighted by "weight_complete_data") having a value in the column shiptype for this MMSI.
\item all_types: all the values of shiptype found in the AIS data for this MMSI.
\item n_type: number of values of shiptype found in the AIS data for this MMSI.
\item type: estimated shiptype of the MMSI, based on the weighted number of points of each shiptype. The shiptype having the greatest score is considered as the real shiptype
\item npoint_type: number of AIS reception having the value "type" (estimated real shiptype) for this MMSI.
\item npoint_type_weighted: number of AIS reception (weighted by "weight_complete_data") having the value "type" (estimated real shiptype) for this MMSI.}
}
\description{
Collect information on the type and length of the vessels to identify the (probable) real type and length, among the errors usually present in the AIS columns. Extract the length and shiptype occurring the most in the MMSI data as the probable real values, and give higher weight to lines with complete data than lines with partial data information. The shiptypes should be checked beforehand to merge the same shiptypes entered as different names. The results should be combined to other AIS data containing this MMSI to confirm the lengths and types returned: analysis showed that daily data are not sufficient to provide reliable values.
}
\examples{
\dontrun{
data("ais")

MMSI_infos(ais_data,
           threshold_length = 475,
           weight_complete_data = 10)}
}
